MscmpSystForms.WebComponents (mscmp_syst_forms v0.1.0)

A set of web components which define the standard application user interface "widgets" from which the application is built.

Containers

mscontainer(assigns)

A generalized container component which provides layouts for and contains other user interface elements.

msdisplay(assigns)

A specialized container component for containing extended text.

msform(assigns)

A specialized container which defines an HTML form.

msinfo(assigns)

A component which provides the capability to include short reference documentation in a form.

mslist(assigns)

A component which defines a list of entries or other components.

mslistitem(assigns)

Creates an list entry within an established mslist/1.

msmodal(assigns)

Generates a modal window.

mssection(assigns)

A container component which typically encapsulates its contents with a border and provides the section a label.

Controls

msbutton(assigns)

Displays a simple button with which the user may interact.

msvalidated_button(assigns)

Provides a three state button where the state is meant to be determined by some external condition, such as a form validation.

Inputs

msinput(assigns)

A component which accepts a wide range of textually oriented user input.

Utility

msfield_errors(assigns)

A utility component which displays an exclamation mark in a circle in the presence of errors.

msicon(assigns)

A small utility component which gives regularized access to "Heroicons".